# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Signaki API
  version: 1.0.0
paths:
...
  /api/v1/comments/{id}/resolve/:
    post:
      operationId: v1_comments_resolve_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ResolveInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResolveInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: ''
...
  /api/v1/photos/{id}/:
    get:
      operationId: v1_photos_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
          description: ''
    patch:
      operationId: v1_photos_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPhotoInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPhotoInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPhotoInput'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: ''
  /api/v1/photos/{id}/favorite/:
    put:
      operationId: v1_photos_favorite_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteContract'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FavoriteContract'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FavoriteContract'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteContract'
          description: ''
  /api/v1/photos/{id}/label/:
    put:
      operationId: v1_photos_label_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LabelInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LabelInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelOutput'
          description: ''
...
  /api/v1/presets/:
    get:
      operationId: v1_presets_list
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PresetContract'
          description: ''
    post:
      operationId: v1_presets_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresetContract'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PresetContract'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PresetContract'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
          description: ''
...
  /api/v1/properties/{id}/decision/:
    get:
      operationId: v1_properties_decision_list
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GalleryDecisionContract'
          description: ''
    post:
      operationId: v1_properties_decision_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DecisionInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryDecisionResult'
          description: ''
...
  /api/v1/properties/{id}/settings/:
    patch:
      operationId: v1_properties_settings_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedGalleryConfiguration'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedGalleryConfiguration'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedGalleryConfiguration'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryConfiguration'
          description: ''
  /api/v1/properties/{id}/upload/:
    post:
      operationId: v1_properties_upload_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UploadInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: ''
...
  /api/v1/versions/{id}/comments/:
    post:
      operationId: v1_versions_comments_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CommentInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CommentInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
          description: ''
  /api/v1/versions/{id}/decision/:
    post:
      operationId: v1_versions_decision_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DecisionInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
          description: ''
  /api/v1/versions/{id}/download/:
    get:
      operationId: v1_versions_download_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
          description: ''
  /api/v1/versions/{id}/information/:
    get:
      operationId: v1_versions_information_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationContract'
          description: ''
components:
  schemas:
    BlankEnum:
      enum:
      - ''
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        version_id:
          type: string
          format: uuid
          readOnly: true
        version_number:
          type: integer
          readOnly: true
        photo_id:
          type: string
          format: uuid
          readOnly: true
        photo_name:
          type: string
          readOnly: true
        property_name:
          type: string
          readOnly: true
        author:
          allOf:
          - $ref: '#/components/schemas/User'
          readOnly: true
        text:
          type: string
          maxLength: 4000
        decision:
          oneOf:
          - $ref: '#/components/schemas/CommentDecisionEnum'
          - $ref: '#/components/schemas/BlankEnum'
        resolved:
          type: boolean
        created:
          type: string
          format: date-time
          readOnly: true
        annotations: {}
        parent:
          type: string
          format: uuid
          nullable: true
      required:
      - author
      - created
      - id
      - photo_id
      - photo_name
      - property_name
      - text
      - version_id
      - version_number
    CommentDecisionEnum:
      enum:
      - approved
      - rejected
      - review
      - in_progress
      type: string
      description: |-
        * `approved` - Approved
        * `rejected` - Rejected — not needed
        * `review` - Review — changes needed
        * `in_progress` - In progress
    CommentInput:
      type: object
      properties:
        annotations: {}
        parent_id:
          type: string
          format: uuid
          nullable: true
        text:
          type: string
          maxLength: 4000
        request_id:
          type: string
          format: uuid
      required:
      - request_id
      - text
...
    DecisionInput:
      type: object
      properties:
        decision:
          $ref: '#/components/schemas/DecisionInputDecisionEnum'
        text:
          type: string
          default: ''
          maxLength: 4000
        expected_revision:
          type: integer
          minimum: 0
        request_id:
          type: string
          format: uuid
      required:
      - decision
      - expected_revision
      - request_id
    DecisionInputDecisionEnum:
      enum:
      - approved
      - rejected
      - review
      - in_progress
      type: string
      description: |-
        * `approved` - approved
        * `rejected` - rejected
        * `review` - review
        * `in_progress` - in_progress
    FavoriteContract:
      type: object
      properties:
        favorite:
          type: boolean
      required:
      - favorite
...
    GalleryConfiguration:
      type: object
      properties:
        view_only:
          type: boolean
        gallery_status:
          type: boolean
        asset_status:
          type: boolean
        favorites:
          type: boolean
        comments:
          type: boolean
        annotations:
          type: boolean
        color_labels:
          type: boolean
        download:
          type: boolean
        upload:
          type: boolean
        file_information:
          type: boolean
        workflow:
          type: boolean
        versioning:
          type: boolean
        watermark:
          type: boolean
        approved_downloads_only:
          type: boolean
        allowed_statuses:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelDefinition'
    GalleryDecisionContract:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        reason:
          type: string
        author:
          type: string
        created:
          type: string
          format: date-time
      required:
      - author
      - created
      - id
      - reason
      - status
    GalleryDecisionResult:
      type: object
      properties:
        status:
          type: string
        revision:
          type: integer
      required:
      - revision
      - status
...
    InformationContract:
      type: object
      properties:
        name:
          type: string
        width:
          type: integer
        height:
          type: integer
        bytes:
          type: integer
        format:
          type: string
        version:
          type: integer
        created:
          type: string
          format: date-time
        kind:
          type: string
      required:
      - bytes
      - created
      - format
      - height
      - kind
      - name
      - version
      - width
    LabelDefinition:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 30
        color:
          type: string
        enabled:
          type: boolean
      required:
      - color
      - enabled
      - id
      - name
    LabelInput:
      type: object
      properties:
        label:
          type: string
      required:
      - label
    LabelOutput:
      type: object
      properties:
        color_label:
          type: string
      required:
      - color_label
...
    PatchedGalleryConfiguration:
      type: object
      properties:
        view_only:
          type: boolean
        gallery_status:
          type: boolean
        asset_status:
          type: boolean
        favorites:
          type: boolean
        comments:
          type: boolean
        annotations:
          type: boolean
        color_labels:
          type: boolean
        download:
          type: boolean
        upload:
          type: boolean
        file_information:
          type: boolean
        workflow:
          type: boolean
        versioning:
          type: boolean
        watermark:
          type: boolean
        approved_downloads_only:
          type: boolean
        allowed_statuses:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelDefinition'
    PatchedPhotoInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 140
        group:
          type: integer
          nullable: true
        position:
          type: integer
          minimum: 0
        hidden:
          type: boolean
...
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        property_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 140
        group_id:
          type: integer
          readOnly: true
          nullable: true
        position:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        hidden:
          type: boolean
        latest:
          allOf:
          - $ref: '#/components/schemas/Version'
          nullable: true
          readOnly: true
        versions:
          type: array
          items:
            $ref: '#/components/schemas/Version'
          readOnly: true
        favorite:
          type: boolean
          readOnly: true
        color_label:
          type: string
          maxLength: 12
      required:
      - favorite
      - group_id
      - id
      - latest
      - name
      - property_id
      - versions
    PhotoResponse:
      type: object
      properties:
        photo:
          $ref: '#/components/schemas/Photo'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
      required:
      - comments
      - photo
    PresetContract:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 80
        config:
          $ref: '#/components/schemas/GalleryConfiguration'
        standard:
          type: boolean
          readOnly: true
      required:
      - config
      - id
      - name
      - standard
...
    ResolveInput:
      type: object
      properties:
        resolved:
          type: boolean
      required:
      - resolved
...
    StatusEnum:
      enum:
      - pending
      - approved
      - rejected
      - review
      - in_progress
      type: string
      description: |-
        * `pending` - Awaiting review
        * `approved` - Approved
        * `rejected` - Rejected — not needed
        * `review` - Review — changes needed
        * `in_progress` - In progress
    UploadInput:
      type: object
      properties:
        group:
          type: integer
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri
          maxItems: 10
          minItems: 1
      required:
      - photos
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
        is_staff:
          type: boolean
          title: Staff status
          description: Designates whether the user can log into this admin site.
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
      required:
      - id
      - username
    Version:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        label:
          type: string
          maxLength: 140
        note:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/StatusEnum'
        created:
          type: string
          format: date-time
          readOnly: true
        review_revision:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        image_url:
          type: string
          readOnly: true
        thumb_url:
          type: string
          readOnly: true
      required:
      - created
      - id
      - image_url
      - label
      - number
      - thumb_url
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid